Removed wihite first from the Board constructor. Refactored moves generation and other minors.
