--- ChessCore/ChessMain.cpp
@@ -1,20 +1,0 @@
-#include "stdafx.h"
-#include "ChessMain.h"
-
-using namespace std;
-using namespace Chess;
-
-ChessMain::ChessMain()
-{
-    cout << "Hello world!"<<endl;
-}
-
-void ChessMain::StartNewGame()
-{
-
-}
-
-ChessMain::~ChessMain()
-{
-
-}
