--- ChessCore/ChessMain.h
@@ -1,14 +1,0 @@
-#pragma once
-#include "Board.h"
-
-namespace Chess
-{
-    class ChessMain
-    {
-    public:
-        ChessMain();
-		void StartNewGame();
-        ~ChessMain();
-    };
-}
-
