+++ ChessCore/Move.h
@@ -3,23 +3,6 @@
 
 namespace Chess
 {
-	struct BinarySerializer
-	{
-		template<class T>
-		static void Serialize(const T & plainObject, std::ostream &ostream)
-		{
-			ostream.write((char*)&plainObject, sizeof(T));
-		}
-
-		template<class T>
-		static T Deserialize(std::istream &istream)
-		{
-			T out;
-			istream.read((char*)&out, sizeof(T));
-			return out;
-		}
-	};
-
 	struct Move
 	{
 		BoardPosition From;
@@ -45,5 +28,10 @@ namespace Chess
 
 		bool IsCapturingMove() const;
 		Move ToMove() const;
+	};
+
+	struct MoveGeneration
+	{
+		static std::vector<Move> GenerateMoves(const Board &board, BoardPosition pieceOffset, EPieceColors side);
 	};
 };
